refactor(cli): split evaluation lifecycle and scoring out of _cli_evaluation
